% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIt-package.R
\docType{package}
\name{FindIt-package}
\alias{FindIt-package}
\alias{CreateWeights}
\alias{CreatelevelIndex}
\alias{lengthSlack}
\alias{Lcombinefunction}
\alias{PsyConstraintCombine}
\alias{CreateANOVAconst}
\alias{Zcombinefunction}
\alias{Glsei}
\alias{CoefExtract}
\alias{rangeCausalANOVAFit}
\alias{CausalANOVAFit}
\alias{stab.CausalANOVA}
\alias{ScreenINT}
\alias{glinternet.cv}
\alias{indTwo2Three}
\alias{NoRegularization}
\alias{Collapsing}
\alias{FullVCOV}
\alias{cluster_se_glm}
\alias{makeallway}
\alias{maketwoway}
\title{FindIt: Finding Heterogeneous Treatment Effects}
\description{
\tabular{ll}{ Package: \tab factorEx\cr Type: \tab Package\cr Version: \tab 1.1.5\cr
Date: \tab 2019-11-19\cr}
}
\references{
Imai, Kosuke and Marc Ratkovic. 2013.  ``Estimating Treatment
Effect Heterogeneity in Randomized Program Evaluation.'' Annals of Applied
Statistics, Vol.7, No.1(March), pp. 443-470.
\url{http://imai.fas.harvard.edu/research/files/svm.pdf}

Egami, Naoki and Kosuke Imai. 2019. Causal Interaction in Factorial Experiments: Application to Conjoint Analysis. Journal of the American Statistical Association, 
Vol.114, No.526 (June), pp. 529–540. \url{http://imai.fas.harvard.edu/research/files/int.pdf}
}
\author{
Naoki Egami, Marc Ratkovic and Kosuke Imai.

Maintainer: Naoki Egami \email{naoki.egami5@gmail.com}
}
\keyword{package}
