\name{full.FindIt}
\alias{full.FindIt}
\title{Generating the full factorial design matrix for a FindIt output}
\description{
  \code{full.FindIt} takes an output object from \code{FindIt} and returns
  the full factorial design matrix for factorial treatments in the data used
  to fit the model.
}
\usage{
full.FindIt(object)
}
\arguments{
  \item{object}{An output object from \code{FindIt}.}
}
\value{
  \item{full}{A full factorial design matrix for factorial treatments in the data used to fit the model.} 
}

\author{Naoki Egami, Marc Ratkovic and Kosuke Imai.}

%\seealso{\link{FindIt}}

\examples{
## See the help page for FindIt() for an example.
	
}