\name{GerberGreen}
\docType{data}
\alias{GerberGreen}
\title{Data from the 1998 New Haven Get-Out-the-Vote Experiment}
\description{This data set contains the full and most recent corrected data from the field experiment analyzed in Gerber and Green (2000).}
\usage{LaLonde}
\format{A data frame consisting of 23 columns and 29,380 observations.}
\source{A data set with columns:
	
persons: Number persons in household

ward: Ward of residence

question: Asked to commit to voting

mailgrp: Sent mail

phonegrp: Phone batch 1

persngrp: Personal contact attempted

appeal: Content of message

  1: Civic Duty
  
  2: Neighborhood Solidarity
  
  3: Close Election

contact: Personal contact occurred

mailings: Number of mailings sent

age: Age of respondent

majorpty: Democratic or Republican

vote96.0: Abstained in 1996

vote96.1: Voted in 1996

mailcall: Phone batch 2

voted98: Voted in 1998

phnscrpt: Script read to phone respondents

  1: Civic-Blood

  2: Civic

  3: Civic or Blood-Civic

  4: Neighbor

  5: Neighbor or Civic-Neighbor

  6: Close

dis.mc: Contacted by phone in batch 2

dis.phn: Contacted by phone in batch 1

phn.c: Contacted by phone

phntrt1: Phone contact attempted (no blood or blood-civic)

phntrt2: Phone contact attempted (no blood)

phn.c1: Contact occurred in phntrt1

phn.c2: Contact occurred in phntrt2

Note: Descriptions come from the description of dataset GerberGreenImai in package Matching.  

}
\references{Gerber, A. S. and Green, D. P. (2000). The effects of canvassing, telephone calls, and direct mail on voter turnout: A field experiment. American Political Science Review 94, 3, 653-663.

Imai, K. (2005). Do get-out-the-vote calls reduce turnout?: The importance of statistical methods for field experiments. American Political Science Review 99, 2, 283-300.

}
\keyword{datasets}