\name{LaLonde}
\docType{data}
\alias{LaLonde}
\title{National Supported Work Study Experimental Data}
\description{This data set gives the outcomes a well as treatment assignments and covariates for the National Supported Work Study, as analyzed in LaLonde (1986).}
\usage{LaLonde}
\format{A data frame consisting of 5 columns (including a treatment assignment vector) and 2787 observations.  }
\source{Data from the National Supported Work Study.  A benchmark matching dataset.  Columns consist of an indicator for whether the observed unit was treated; an indicator for whether the individual received the treatment; age in years; schooling in years; indicators for black and Hispanic; an indicator for marriage status, one of married; an indicator for no high school degree; reported earnings in 1975 and 1978; and extrapolation weights to the 1978 Panel Study for Income Dynamics dataset.  1974 and 1975 earnings are pre-treatment.  1978 earnings is taken as the outcome variable.}
\references{LaLonde, R.J. (1986). Evaluating the econometric evaulations of training programs with experimental data. American Economic Review 76, 4, 604-620.}
\keyword{datasets}