% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CausalANOVA.R
\name{plot.CausalANOVA}
\alias{plot.CausalANOVA}
\title{Plotting CausalANOVA}
\usage{
\method{plot}{CausalANOVA}(
  x,
  fac.name,
  treat.ind = 1,
  type = "ConditionalEffect",
  space = 15,
  xlim,
  ...
)
}
\arguments{
\item{x}{An output from \code{CausalANOVA}}

\item{fac.name}{Factor names to plot. Length should be 2.}

\item{treat.ind}{Which factor serves as the main treatment. Should be 1 (the first element of \code{fac.name}) or 2 (the second element of \code{fac.name}).}

\item{type}{What types of effects to plot. Should be one of \code{AME}, \code{AMIE} and \code{ConditionalEffect}.}

\item{space}{Space on the left side of the plot.}

\item{xlim}{Range for the x-axis}

\item{...}{Other graphical parameters}
}
\description{
Plotting CausalANOVA
}
