% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CausalANOVA.R
\name{summary.CausalANOVA}
\alias{summary.CausalANOVA}
\title{Summarizing CausalANOVA output}
\usage{
\method{summary}{CausalANOVA}(object, digit = 4, verbose = TRUE, verbose.full = TRUE, ...)
}
\arguments{
\item{object}{An object from \code{CausalANOVA}}

\item{digit}{the number of digits}

\item{verbose}{report additional summary results}

\item{verbose.full}{report full summary results}

\item{...}{Other parameters}
}
\description{
Summarizing CausalANOVA output
}
