% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FindIt.R
\name{summary.FindIt}
\alias{summary.FindIt}
\title{Summarizing FindIt output}
\usage{
\method{summary}{FindIt}(object, ...)
}
\arguments{
\item{object}{An object from \code{FindIt}}

\item{...}{Other parameters}
}
\description{
Summarizing FindIt output
}
