% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIt-package.R
\docType{data}
\name{Carlson}
\alias{Carlson}
\title{Data from conjoint analysis in Carlson (2015).}
\format{A data frame consisting of 7 columns (including a treatment
assignment vector) and 3232 observations.

\tabular{llll}{ outcome \tab integer \tab whether a profile is chosen \tab
0,1 \cr newRecordF \tab factor \tab record as a politician \tab 7 levels \cr
promise \tab factor \tab platform \tab 3 levels (job, clinic, education) \cr
coeth_voting \tab factor \tab whether a profile is coethnic to a respodent
\tab Yes, No \cr Degree \tab factor \tab job whether a profile has relevant
degrees \tab 4 Yes, No \cr }}
\source{
Data from Carlson (2015).
}
\description{
This data set gives the outcomes a well as treatment assignments of the
conjoint analysis in Carlson (2015). Please Carlson (2015) and Egami and
Imai (2016+) for more details.
}
\references{
Carlson, E. 2015. ``Ethnic voting and accountability in africa:
A choice experiment in uganda.'' World Politics 67, 02, 353--385.
}
\keyword{datasets}
