\name{FindIt-internal}
\title{Internal FindIt functions}
\alias{CreateWeights}
\alias{CreatelevelIndex}
\alias{lengthSlack}
\alias{Lcombinefunction}
\alias{PsyConstraintCombine}
\alias{CreateWeights}
\alias{CreateANOVAconst}
\alias{Zcombinefunction}
\alias{Glsei}
\alias{CoefExtract}
\alias{rangeCausalANOVAFit}
\alias{CausalANOVAFit}
\alias{stab.CausalANOVA}
\alias{AMIEFit}
\alias{makeallway}
\alias{maketwoway}
\description{Internal FindIt functions}
\author{Naoki Egami}
\details{These are not intended for use by users.}
\keyword{internal}