% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindIt-package.R
\docType{data}
\name{GerberGreen}
\alias{GerberGreen}
\title{Data from the 1998 New Haven Get-Out-the-Vote Experiment}
\format{A data frame consisting of 9 columns and 29,380 observations.

\tabular{llll}{ voted98 \tab integer \tab voted in 1998 \tab 0,1 \cr
persngrp \tab factor \tab personal contact attempted \tab 0,1 \cr phnscrpt
\tab factor \tab script read to phone respondents \tab 7 levels \cr mailings
\tab factor \tab number of mailings sent \tab 0 - 3 \cr appeal \tab factor
\tab content of message \tab 3 levels \cr age \tab integer \tab age of
respondent \tab \cr majorpty \tab factor \tab Democratic or Republican \tab
\cr voted96.1 \tab factor \tab voted in 1996 \tab 0,1 \cr voted96.0 \tab
factor \tab abstained in 1996 \tab 0,1 \cr }

Note: The levels of phnscrpt and appeal are follows.

phnscrpt: Script read to phone respondents \tabular{ll}{ 0 \tab No phone \cr
1 \tab Civic-Blood \cr 2 \tab Civic \cr 3 \tab Civic or Blood-Civic \cr 4
\tab Neighbor \cr 5 \tab Neighbor or Civic-Neighbor \cr 6 \tab Close }

appeal: Content of message \tabular{ll}{ 1 \tab Civic Duty \cr 2 \tab
Neighborhood Solidarity \cr 3 \tab Close Election }}
\description{
This data set contains the most recent corrected data from the field
experiment analyzed in Gerber and Green (2000).
}
\references{
Gerber, A. S. and Green, D. P. 2000 . ``The effects of
canvassing, telephone calls, and direct mail on voter turnout: A field
experiment.'' American Political Science Review, Vol.94, No.3, pp. 653-663.

Imai, K. 2005 . ``Do get-out-the-vote calls reduce turnout?: The importance
of statistical methods for field experiments."" American Political Science
Review, Vol.99, No.2, pp. 283-300.
}
\keyword{datasets}
